Separated option menu item implementations into a new source file, optionmenuitems.cpp.
