+++ src/CMakeLists.txt
@@ -981,6 +981,7 @@ set( FASTMATH_PCH_SOURCES
 	menu/menuinput.cpp
 	menu/messagebox.cpp
 	menu/optionmenu.cpp
+	menu/optionmenuitems.cpp
 	menu/playerdisplay.cpp
 	menu/playermenu.cpp
 	menu/readthis.cpp
